Make the default buffer size 512 samples.
Experiments with filling up the write pipe suggest that the total
playback queue capacity of the QA401 is 1024 samples, and that seems
to be true of the record queue as well. Using the queue size as the
ASIO buffer size is a bad idea because it means reads will not
happen in time to avoid overflowing the QA401 queue.