Do not try to finish writing before starting the QA401.
The previous code made it impossible to use a buffer size that is
larger than half the QA401 output queue, because we were trying to
get two buffers to fit in there before starting the QA401. With this
change, we will start the hardware as soon as at least one output
buffer made it.