Do the first Reset() call when we are about to start streaming.
This seems somewhat cleaner, as this way we don't alter the state of
the QA401 until we actually start using it. Besides, this will help
with managing the state of register 5, especially since the sample
rate can only be changed when the QA401 is reset, but we don't know
the sample rate yet at ASIO init time.

See #2.