Address QA401 start issues by rewriting most of the streaming logic.
The main goal of this commit is to work around #9 by getting rid of
QA401::Start(), which is now folded into Reset() so that the QA401
does not get to spend any significant amount of time in a "weird"
"DC party" state.

This required a rethink of the priming logic and of the streaming
loop. The good news is, the new logic is arguably cleaner and
simpler, and it has the nice side effect of taking care of #5
without the need for additional workarounds.

Fixes #9. Fixes #5.