Rewrite the streaming logic
This is a complete rewrite of the core ASIO401 streaming logic to make
it simpler, more robust, more performant, and more correct with
regard to the ASIO APIs semantics.

By far the biggest change is the use of double-buffering on the QA40x
side to avoid timing pressure on the worker thread when an I/O completes
and the next one has to be issued.

Other improvements include better priming logic that properly waits
until priming is done before starting the hardware, and avoids
unnecessary I/O by combining initial padding with the first buffer.