Enforce write granularity
I recently discovered that the output of the DAC is garbled if we write
a buffer whose size is not a multiple of 64 frames (32 for the QA401).
Let's use that information to prevent the user from shooting themselves
in the foot.

I have empirically verified that the input (ADC) side does not suffer
from this problem, and doesn't seem to have any granularity constraints.